rename reportviews to views and hypergraphviews to globalviews
