--- docs/source/api/core/xgi.core.hypergraphviews.rst
@@ -1,10 +1,0 @@
-﻿xgi.core.hypergraphviews
-========================
-
-.. currentmodule:: xgi.core.hypergraphviews
-
-.. automodule:: xgi.core.hypergraphviews
-
-   .. rubric:: Functions
-
-   .. autofunction:: subhypergraph
