+++ tests/core/test_dihypergraph.py
