+++ tests/core/test_diviews.py
