+++ tests/core/test_function.py
