+++ tests/core/test_simplicialcomplex.py
