+++ tests/core/test_views.py
