+++ xgi/core/__init__.py
@@ -1,5 +1,5 @@
 from .hypergraph import Hypergraph
 from .dihypergraph import DiHypergraph
 from .function import *
-from .hypergraphviews import subhypergraph
+from .globalviews import subhypergraph
 from .simplicialcomplex import SimplicialComplex
