+++ xgi/core/dihypergraph.py
@@ -11,7 +11,7 @@ from warnings import warn
 
 from ..exception import XGIError
 from ..utils import IDDict, update_uid_counter
-from .direportviews import DiEdgeView, DiNodeView
+from .diviews import DiEdgeView, DiNodeView
 
 __all__ = ["DiHypergraph"]
