+++ xgi/core/diviews.py
