+++ xgi/core/globalviews.py
@@ -7,7 +7,7 @@ re-add. This module provides those graph views.
 The resulting views are essentially read-only graphs that report data from the original
 graph object.
 
-Note: Since hypergraphviews look like hypergraphs, one can end up with
+Note: Since globalviews look like hypergraphs, one can end up with
 view-of-view-of-view chains. Be careful with chains because they become very slow with
 about 15 nested views. Often it is easiest to use .copy() to avoid chains.
