+++ xgi/core/views.py
